RabbtMQ

+++ vmall/mall-service/src/main/java/vn/vti/dtn2501/mall/payload/response/CreateCartItemResponse.java
@@ -13,6 +13,6 @@ public class CreateCartItemResponse {
     private Long cartId;
     private Long productId;
     private String productName;
-    private BigDecimal productPrice;
+    private Double productPrice;
     private Long quantity;
 }
